Move qrcode resolvers into dedicated modules
