+++ back/boxtribute_server/graph_ql/resolvers.py
@@ -1,33 +1,11 @@
 """GraphQL resolver functionality"""
-from ariadne import MutationType, QueryType, convert_kwargs_to_snake_case
+from ariadne import MutationType, QueryType
 
 from ..authz import authorize
-from ..models.crud import create_qr_code
-from ..models.definitions.box import Box
-from ..models.definitions.location import Location
-from ..models.definitions.qr_code import QrCode
-from .bindables import beneficiary, qr_code, unboxed_items_collection
+from .bindables import beneficiary, unboxed_items_collection
 
 query = QueryType()
 mutation = MutationType()
-
-
-@query.field("qrExists")
-@convert_kwargs_to_snake_case
-def resolve_qr_exists(*_, qr_code):
-    authorize(permission="qr:read")
-    try:
-        QrCode.get_id_from_code(qr_code)
-    except QrCode.DoesNotExist:
-        return False
-    return True
-
-
-@query.field("qrCode")
-@convert_kwargs_to_snake_case
-def resolve_qr_code(obj, _, qr_code=None):
-    authorize(permission="qr:read")
-    return obj.qr_code if qr_code is None else QrCode.get(QrCode.code == qr_code)
 
 
 @unboxed_items_collection.field("product")
@@ -35,21 +13,7 @@ def resolve_box_product(obj, info):
     return info.context["product_loader"].load(obj.product_id)
 
 
-@mutation.field("createQrCode")
-@convert_kwargs_to_snake_case
-def resolve_create_qr_code(*_, box_label_identifier=None):
-    authorize(permission="qr:create")
-    return create_qr_code(box_label_identifier=box_label_identifier)
-
-
 @beneficiary.field("base")
 def resolve_resource_base(obj, _):
     authorize(permission="base:read", base_id=obj.base_id)
     return obj.base
-
-
-@qr_code.field("box")
-def resolve_qr_code_box(qr_code_obj, _):
-    box = Box.select().join(Location).where(Box.qr_code == qr_code_obj.id).get()
-    authorize(permission="stock:read", base_id=box.location.base_id)
-    return box
